User Repository with business rules

+++ Assignment3.Entities/UserRepository.cs
@@ -16,6 +16,11 @@ public class UserRepository : IUserRepository
         entity.Email = user.Email;
         entity.Name = user.Name;
 
+        var userExists = _context.Users.FirstOrDefault(u => u.Email == user.Email) != null ? true : false;
+        if(userExists) 
+        {
+            return (Response.Conflict, 0);
+        }
         _context.Users.Add(entity);
         _context.SaveChanges();
 
@@ -28,11 +33,15 @@ public class UserRepository : IUserRepository
     public Response Delete(int userId, bool force = false)
     {
         var user = _context.Users.FirstOrDefault(u => u.Id == userId);
-
         if(user == null)
         {
             return Response.NotFound;
         } 
+        bool AssignedToTask = false;
+        if(AssignedToTask && !force) 
+        {
+            return Response.Conflict;
+        }
         _context.Users.Remove(user!);
         return Response.Deleted;
     }
